refactor #36: useLocationMutation에 있던 훅들 한 곳으로 모으기
