+++ src/components/LocationButton.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react';
 import { styled } from 'styled-components';
-import { useSelectLocation } from '../queries/useLocationMutation';
+import { useSelectLocation } from '../queries/useLocationQuery';
 import { useLocationStore } from '../stores/useLocationStore';
 import { Button } from './Button';
 import { Icon } from './icon/Icon';
